Modularize tests using shared file with a common fixture
